# Docs/openapi_real.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alling B2B API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/checkout/": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Procesar Checkout",
        "description": "Procesa un Formato Único en estado COTIZACION, calculando envío\ny generando una preferencia de pago.",
        "operationId": "procesar_checkout_checkout__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/checkout/{fu_id}/status": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "summary": "Get Checkout Status",
        "description": "Retorna el estado de un FSM de manera ligera.",
        "operationId": "get_checkout_status_checkout__fu_id__status_get",
        "parameters": [
          {
            "name": "fu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutStatusSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/consultas": {
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Listar Consultas",
        "description": "RF-CON-001: Lista FormatosUnicos en estado CONSULTA.\nRF-CON-004: Filtra por asignación (OPS-CON-004).\nOPS-CON-001: SELLER ve todas las consultas (asignadas y sin asignar).",
        "operationId": "listar_consultas_consultas_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assigned_to_me",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "True: mis consultas. False: sin asignar. None: todas",
              "title": "Assigned To Me"
            },
            "description": "True: mis consultas. False: sin asignar. None: todas"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaListSchema"
                  },
                  "title": "Response Listar Consultas Consultas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consultas/{fu_id}/tomar": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Tomar Consulta",
        "description": "RF-CON-002: Asigna la consulta al SELLER actual.\nRN-CON-001: solo un SELLER puede estar asignado a la vez.\nHTTP 409 si ya está asignada a otro SELLER (optimistic lock).\nEVT-CON-001: ConsultaAsignada.",
        "operationId": "tomar_consulta_consultas__fu_id__tomar_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fu Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TomarConsultaResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/consultas/{fu_id}/responder": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Responder Consulta",
        "description": "RF-CON-003: Responde la consulta y transiciona FU a RESUELTA (FU-T-05).\nRN-CON-002: Solo el SELLER asignado puede responder.\nconsultant_note no puede estar vacía (validado por Pydantic min_length=1).\nEVT-FU-012: ConsultaResuelta.",
        "operationId": "responder_consulta_consultas__fu_id__responder_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fu Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponderConsultaSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponderConsultaResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutRequestSchema": {
        "properties": {
          "fu_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fu Id"
          },
          "billing_id": {
            "type": "string",
            "title": "Billing Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "fu_id",
          "billing_id",
          "address"
        ],
        "title": "CheckoutRequestSchema"
      },
      "CheckoutResponseSchema": {
        "properties": {
          "fu_id": {
            "type": "string",
            "title": "Fu Id"
          },
          "payment_url": {
            "type": "string",
            "title": "Payment Url"
          },
          "shipping_cost": {
            "type": "string",
            "title": "Shipping Cost"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "order_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Token"
          }
        },
        "type": "object",
        "required": [
          "fu_id",
          "payment_url",
          "shipping_cost",
          "idempotency_key"
        ],
        "title": "CheckoutResponseSchema"
      },
      "CheckoutStatusSchema": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "CheckoutStatusSchema"
      },
      "ConsultaListSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "assigned_seller_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Seller Id"
          },
          "consultant_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consultant Note"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "title": "ConsultaListSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponderConsultaResponseSchema": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fu_id": {
            "type": "string",
            "title": "Fu Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "consultant_note": {
            "type": "string",
            "title": "Consultant Note"
          }
        },
        "type": "object",
        "required": [
          "message",
          "fu_id",
          "state",
          "consultant_note"
        ],
        "title": "ResponderConsultaResponseSchema"
      },
      "ResponderConsultaSchema": {
        "properties": {
          "consultant_note": {
            "type": "string",
            "minLength": 1,
            "title": "Consultant Note",
            "description": "Nota de asesoría. No puede estar vacía (RN-CON-002)"
          }
        },
        "type": "object",
        "required": [
          "consultant_note"
        ],
        "title": "ResponderConsultaSchema"
      },
...
      "TomarConsultaResponseSchema": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fu_id": {
            "type": "string",
            "title": "Fu Id"
          },
          "assigned_seller_id": {
            "type": "string",
            "title": "Assigned Seller Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "fu_id",
          "assigned_seller_id"
        ],
        "title": "TomarConsultaResponseSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}